update walk-in booking success

+++ resources/views/backoffice/rooms.blade.php
@@ -59,8 +59,10 @@
                                                 class="bi bi-pencil-square"></i></button>
                                         <button onclick="getGallery('{{ $room->id }}')" class="btn-modal btn btn-info shadow-none" data-bs-toggle="modal"
                                             data-bs-target="#images-s"><i class="bi bi-images"></i></button>
-                                        <button class="btn btn-danger shadow-none" onclick="deleteRoom({{ $room->id }})"><i
-                                                class="bi bi-trash-fill"></i></button>
+                                        @if ($shareUser->admin_role === 'แอดมินสูงสุด')
+                                            <button class="btn btn-danger shadow-none" onclick="deleteRoom({{ $room->id }})"><i
+                                                    class="bi bi-trash-fill"></i></button>
+                                        @endif
                                     </div>
                                 </td>
                             </tr>
@@ -324,6 +326,7 @@
             </form>
         </div>
     </div>
+
 @endsection
 
 @section('script')
